# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KinConnect API
  description: |
    KinConnect is a comprehensive medication management platform that helps patients and caregivers 
    manage medications, track adherence, and coordinate care within family groups.
    
    ## Features
    - **Medication Management**: Create, update, and track medications with reminders
    - **Drug Search**: Search RxNorm database for drug information, interactions, and images
    - **Family Groups**: Invite and manage family members with role-based access
    - **Patient Profiles**: Comprehensive patient health information management
    - **Audit Logging**: Track all access and modifications for compliance
    
    ## Authentication
    All API endpoints require Firebase Authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Rate Limiting
    API requests are limited to 100 requests per 15-minute window per IP address.
    
    ## Versioning
    Current API version: v1
    
  version: 1.0.0
  contact:
    name: KinConnect Support
    email: support@kinconnect.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5001/{project-id}/us-central1/api
    description: Local Development Server
    variables:
      project-id:
        default: kinconnect-dev
        description: Firebase Project ID
  - url: https://us-central1-{project-id}.cloudfunctions.net/api
...
    variables:
...
security:
  - bearerAuth: []
paths:
...
  /medications:
    get:
      tags:
        - Medications
      summary: List all medications
      description: Retrieves all medications for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Medications retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Medication'
              examples:
                success:
                  value:
                    success: true
                    data:
                      - id: "med123"
                        patientId: "user123"
                        name: "Lisinopril"
                        genericName: "Lisinopril"
                        brandName: "Prinivil"
                        rxcui: "104377"
                        dosage: "10mg"
                        frequency: "Once daily"
                        instructions: "Take with food"
                        prescribedBy: "Dr. Smith"
                        prescribedDate: "2024-01-01T00:00:00Z"
                        isActive: true
                        createdAt: "2024-01-01T00:00:00Z"
                        updatedAt: "2024-01-01T00:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Medications
      summary: Create a new medication
      description: Creates a new medication record for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationInput'
            examples:
              basic:
                value:
                  name: "Lisinopril"
                  dosage: "10mg"
                  frequency: "Once daily"
                  instructions: "Take with food"
                  prescribedBy: "Dr. Smith"
                  isActive: true
              complete:
                value:
                  name: "Lisinopril"
                  genericName: "Lisinopril"
                  brandName: "Prinivil"
                  rxcui: "104377"
                  ndc: "0378-0710-93"
                  dosage: "10mg"
                  strength: "10mg"
                  dosageForm: "tablet"
                  frequency: "Once daily"
                  route: "oral"
                  instructions: "Take with food in the morning"
                  prescribedBy: "Dr. Smith"
                  prescribedDate: "2024-01-01"
                  startDate: "2024-01-01"
                  isActive: true
                  isPRN: false
                  pharmacy: "CVS Pharmacy"
                  prescriptionNumber: "RX123456"
                  refillsRemaining: 3
                  sideEffects: ["Dizziness", "Dry cough"]
                  notes: "Monitor blood pressure weekly"
      responses:
        '201':
          description: Medication created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Medication'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /medications/{id}:
    put:
      tags:
        - Medications
      summary: Update a medication
      description: Updates an existing medication record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Medication ID
          schema:
            type: string
          example: "med123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationInput'
      responses:
        '200':
          description: Medication updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Medication'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Medications
      summary: Delete a medication
      description: Deletes a medication record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Medication ID
          schema:
            type: string
          example: "med123"
      responses:
        '200':
          description: Medication deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "Medication deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /medications/{medicationId}/reminders:
    get:
      tags:
        - Medications
      summary: Get medication reminders
      description: Retrieves all reminders for a specific medication
      security:
        - bearerAuth: []
      parameters:
        - name: medicationId
          in: path
          required: true
          description: Medication ID
          schema:
            type: string
          example: "med123"
      responses:
        '200':
          description: Reminders retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/MedicationReminder'
              examples:
                success:
                  value:
                    success: true
                    data:
                      - id: "reminder123"
                        medicationId: "med123"
                        patientId: "user123"
                        reminderTime: "08:00"
                        days: ["monday", "tuesday", "wednesday", "thursday", "friday"]
                        isActive: true
                        createdAt: "2024-01-01T00:00:00Z"
                        updatedAt: "2024-01-01T00:00:00Z"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Medications
      summary: Create medication reminder
      description: Creates a new reminder for a medication
      security:
        - bearerAuth: []
      parameters:
        - name: medicationId
          in: path
          required: true
          description: Medication ID
          schema:
            type: string
          example: "med123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationReminderInput'
            examples:
              daily:
                value:
                  reminderTime: "08:00"
                  days: ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
                  isActive: true
              weekdays:
                value:
                  reminderTime: "09:00"
                  days: ["monday", "tuesday", "wednesday", "thursday", "friday"]
                  isActive: true
      responses:
        '201':
          description: Reminder created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MedicationReminder'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /medications/reminders/{reminderId}:
    put:
      tags:
        - Medications
      summary: Update medication reminder
      description: Updates an existing medication reminder
      security:
        - bearerAuth: []
      parameters:
        - name: reminderId
          in: path
          required: true
          description: Reminder ID
          schema:
            type: string
          example: "reminder123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationReminderInput'
      responses:
        '200':
          description: Reminder updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MedicationReminder'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags:
        - Medications
      summary: Delete medication reminder
      description: Deletes a medication reminder
      security:
        - bearerAuth: []
      parameters:
        - name: reminderId
          in: path
          required: true
          description: Reminder ID
          schema:
            type: string
          example: "reminder123"
      responses:
        '200':
          description: Reminder deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "Reminder deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== DRUG SEARCH ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained from Firebase Authentication
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message about the result
        error:
          type: string
          description: Error message if request failed
        data:
          description: Response data (type varies by endpoint)

    # ==================== USER & PATIENT ====================
...
    Medication:
      type: object
      required:
        - id
        - patientId
        - name
        - dosage
        - frequency
        - instructions
        - prescribedBy
        - prescribedDate
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Medication ID
        patientId:
          type: string
          description: Patient ID this medication belongs to
        name:
          type: string
          description: Medication name
        genericName:
          type: string
          description: Generic name of the medication
        brandName:
          type: string
          description: Brand name of the medication
        rxcui:
          type: string
          description: RxNorm Concept Unique Identifier
        ndc:
          type: string
          description: National Drug Code
        dosage:
          type: string
          description: Dosage amount (e.g., "10mg")
        strength:
          type: string
          description: Medication strength
        dosageForm:
          type: string
          description: Form of medication (tablet, capsule, liquid, etc.)
        frequency:
          type: string
          description: How often to take (e.g., "Once daily")
        route:
          type: string
          description: Route of administration (oral, topical, injection, etc.)
        instructions:
          type: string
          description: Specific instructions for taking the medication
        prescribedBy:
          type: string
          description: Name of prescribing physician
        prescribedDate:
          type: string
          format: date-time
          description: Date medication was prescribed
        startDate:
          type: string
          format: date-time
          description: Date to start taking medication
        endDate:
          type: string
          format: date-time
          description: Date to stop taking medication
        isActive:
          type: boolean
          description: Whether medication is currently active
        isPRN:
          type: boolean
          description: Whether medication is taken "as needed"
        maxDailyDose:
          type: string
          description: Maximum daily dose
        sideEffects:
          type: array
          items:
            type: string
          description: Known side effects
        notes:
          type: string
          description: Additional notes
        pharmacy:
          type: string
          description: Pharmacy name
        prescriptionNumber:
          type: string
          description: Prescription number
        refillsRemaining:
          type: integer
          description: Number of refills remaining
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MedicationInput:
      type: object
      required:
        - name
        - dosage
        - frequency
        - instructions
        - prescribedBy
        - isActive
      properties:
        name:
          type: string
        genericName:
          type: string
        brandName:
          type: string
        rxcui:
          type: string
        ndc:
          type: string
        dosage:
          type: string
        strength:
          type: string
        dosageForm:
          type: string
        frequency:
          type: string
        route:
          type: string
        instructions:
          type: string
        prescribedBy:
          type: string
        prescribedDate:
          type: string
          format: date
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        isPRN:
          type: boolean
        maxDailyDose:
          type: string
        sideEffects:
          type: array
          items:
            type: string
        notes:
          type: string
        pharmacy:
          type: string
        prescriptionNumber:
          type: string
        refillsRemaining:
          type: integer
    MedicationReminder:
      type: object
      required:
        - id
        - medicationId
        - patientId
        - reminderTime
        - days
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Reminder ID
        medicationId:
          type: string
          description: Associated medication ID
        patientId:
          type: string
          description: Associated patient ID
        reminderTime:
          type: string
          pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
          description: Time of day for reminder (HH:MM format)
          example: "08:00"
        days:
          type: array
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
          description: Days of the week for reminder
        isActive:
          type: boolean
          description: Whether reminder is active
        lastNotified:
          type: string
          format: date-time
          description: Last time notification was sent
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MedicationReminderInput:
      type: object
      required:
        - reminderTime
        - days
        - isActive
      properties:
        reminderTime:
          type: string
          pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
          example: "08:00"
        days:
          type: array
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        isActive:
          type: boolean

    # ==================== DRUG SEARCH ====================
...
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            missing_token:
              value:
                success: false
                error: "Authorization header required"
            invalid_token:
              value:
                success: false
                error: "Invalid or expired token"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            not_found:
              value:
                success: false
                error: "Resource not found"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            server_error:
              value:
                success: false
                error: "Internal server error"